[UPDATE] Updated service management

+++ ISS/Projekatsiit2024/src/main/java/rs/ac/uns/ftn/asd/Projekatsiit2024/controller/OfferCategoryController.java
@@ -6,6 +6,7 @@ import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,6 +22,7 @@ import rs.ac.uns.ftn.asd.Projekatsiit2024.dto.offerCategory.MinimalOfferCategory
 import rs.ac.uns.ftn.asd.Projekatsiit2024.dto.offerCategory.PostOfferCategoryDTO;
 import rs.ac.uns.ftn.asd.Projekatsiit2024.dto.offerCategory.PutOfferCategoryDTO;
 import rs.ac.uns.ftn.asd.Projekatsiit2024.exception.event.EventTypeValidationException;
+import rs.ac.uns.ftn.asd.Projekatsiit2024.model.auth.UserPrincipal;
 import rs.ac.uns.ftn.asd.Projekatsiit2024.model.offer.OfferCategory;
 import rs.ac.uns.ftn.asd.Projekatsiit2024.service.OfferCategoryService;
 
@@ -33,8 +35,12 @@ public class OfferCategoryController {
 	@GetMapping
 	public ResponseEntity<List<MinimalOfferCategoryDTO>> getCategories(
 			@RequestParam(name = "isAccepted", required = false) Boolean isAccepted,
-			@RequestParam(name = "isEnabled", required = false) Boolean isEnabled
+			@RequestParam(name = "isEnabled", required = false) Boolean isEnabled,
+			@AuthenticationPrincipal UserPrincipal userPrincipal
 			) {
+		if(!userPrincipal.getUser().getRole().getName().equals("ADMIN_ROLE"))
+			return ResponseEntity.status(404).body(null);
+		
 		List<OfferCategory> ocs = offerCategoryService.getOffers(isAccepted, isEnabled);
 		List<MinimalOfferCategoryDTO> miniOcs = ocs.stream().map(oc -> new MinimalOfferCategoryDTO(oc)).toList();
 		return ResponseEntity.ok(miniOcs);
@@ -57,8 +63,12 @@ public class OfferCategoryController {
 	}
 	
 	@PostMapping
-	public ResponseEntity createNewCategory(@RequestBody PostOfferCategoryDTO data) {
+	public ResponseEntity createNewCategory(@RequestBody PostOfferCategoryDTO data,
+			@AuthenticationPrincipal UserPrincipal userPrincipal) {
 		try {
+			if(!userPrincipal.getUser().getRole().getName().equals("ADMIN_ROLE"))
+				return ResponseEntity.status(404).body(null);
+			
 			OfferCategory oc = offerCategoryService.createAndFlush(data.name, data.description);
 			return ResponseEntity.status(201).body(new MinimalOfferCategoryDTO(oc));
 		} catch (IllegalArgumentException e) {
@@ -67,8 +77,12 @@ public class OfferCategoryController {
 	}
 	
 	@PutMapping("/{id}")
-	public ResponseEntity editExistingCategory(@PathVariable Integer id, @RequestBody PutOfferCategoryDTO data) {
+	public ResponseEntity editExistingCategory(@PathVariable Integer id, @RequestBody PutOfferCategoryDTO data,
+			@AuthenticationPrincipal UserPrincipal userPrincipal) {
 		try {
+			if(!userPrincipal.getUser().getRole().getName().equals("ADMIN_ROLE"))
+				return ResponseEntity.status(404).body(null);
+			
 			offerCategoryService.editCategory(id, data.name, data.description, data.isEnabled);
 			return ResponseEntity.noContent().build();
 		} catch (IllegalArgumentException e) {
@@ -77,9 +91,12 @@ public class OfferCategoryController {
 	}
 	
 	@PutMapping("/pending/{id}")
-	public ResponseEntity handleCategorySuggestion(@PathVariable Integer id, @RequestBody HandleSuggestionDTO data) {
-		//TODO: Admin validation
+	public ResponseEntity handleCategorySuggestion(@PathVariable Integer id, @RequestBody HandleSuggestionDTO data,
+			@AuthenticationPrincipal UserPrincipal userPrincipal) {
 		try {
+			if(!userPrincipal.getUser().getRole().getName().equals("ADMIN_ROLE"))
+				return ResponseEntity.status(404).body(null);
+			
 			OfferCategory oc;
 			if(data.isAccepted)
 				oc = offerCategoryService.acceptSuggestion(id, data.name, data.description);
@@ -92,8 +109,12 @@ public class OfferCategoryController {
 	}
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity deleteCategory(@PathVariable Integer id) {
+	public ResponseEntity deleteCategory(@PathVariable Integer id,
+			@AuthenticationPrincipal UserPrincipal userPrincipal) {
 		try {
+			if(!userPrincipal.getUser().getRole().getName().equals("ADMIN_ROLE"))
+				return ResponseEntity.status(404).body(null);
+			
 			offerCategoryService.deleteCategory(id);
 			return ResponseEntity.noContent().build();
 		} catch (IllegalArgumentException e) {
